splitted new functions out in separate C++ file data_utilities.cpp
